refactor: massively overhaul project structure
